app/requirements.txt
flask
pytesseract
Pillow
gunicorn